requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.2
rich>=13.7.0
typer>=0.9.0

# HTTP / Async
httpx>=0.26.0
requests>=2.31.0
tenacity>=8.2.3

# Data / Math
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Scheduling
APScheduler>=3.10.4

# Geocoding
geopy>=2.4.1

# Telegram
python-telegram-bot>=20.7

# Polymarket
py-clob-client>=0.18.0

# Crypto / Auth (required by py-clob-client)
eth-account>=0.10.0
web3>=6.11.0

# Storage
tinydb>=4.8.0

# Plotting (optional, for distribution charts)
matplotlib>=3.8.0

# Timezone
pytz>=2024.1
tzdata>=2024.1

# YAML
PyYAML>=6.0.1
